pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.33.0",
    "graphviz>=0.21",
    "ipykernel>=6.30.1",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.6",
    "matplotlib>=3.10.6",
    "networkx>=3.5",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]
